Create a header file which provides the build information
Doing this way, we don't need to modify the sources under version control!
As the sed modification of the source file, the generation of the new
header file is _not_ phony.
Also removed the sed command, which is not used anymore.

buildinfo.h depends on all source files

This way the buildinfo.h is rebuild each time when the sources have
changed.

Delete the buildinfo.h file on make goal clean

Use precompiler macros for the BUILD_TIMESTAMP in case of MSVC builds,
which can not create buildinfo.h.

Signed-off-by: Juergen Kosel <juergen.kosel@softing.com>