Remove Log_levels global definition.
The -fno-common flag indicates the Log_levels is defined in many
places.

First, it's not required to have a global Log_levels, just the
enum type LOG_LEVELS. The problem of define a global variable in
many modules is that the linker may choose the wrong one. This
causes very hard-to-find linker bugs.

Second, the Log_levels is not extern and waste binary space.

Signed-off-by: Guilherme Maciel Ferreira <guilherme.maciel.ferreira@gmail.com>