build: various warning fixes
This patch fixes various warnings produced during the build the most
common warning, is the truncation of "size_t" (usually unsigned long) to
(int).  This patch explicitly casts the result to "int" in these cases
to supress the warning.  Additionally, this patch includes a fix to an
internal library load function to take a "const char *" parameter
instead of a "char *" parameter.

Signed-off-by: Keith Holman <keith.holman@windriver.com>