Update Makefile
When the library is built in a Yocto environment, make is called with the -j option to parallelize the build.
In this case, the VersionInfo.h creation with sed can be launched at the same time than the mkdir target and sometimes sed is executed before the actual directory creation.

Here is the error log:
$ sed -e "s/@CLIENT_VERSION@/1.2.1/g" -e "s/@BUILD_TIMESTAMP@/Thu Aug  2 10:21:03 UTC 2018/g" src/VersionInfo.h.in > build/VersionInfo.h
cannot create build/VersionInfo.h: Directory nonexistent

The idea of this commit is to ensure that the directory is created before calling sed by adding a call to mkdir in the VersionInfo.h's Makefile recipe.

Signed-off-by: aizelauna <aizelauna@users.noreply.github.com>