Add disconnect option initializer for mqttv5
New fields where added to the MQTTAsync_disconnectOptions structure for
MQTTv5 support. The macro provided to initialize this macro, does not
included the new fields. Depending on compilation options it may cause a
warning similar to: 'warning: missing initializer for field
‘onSuccess5’".

Add a MQTTAsync_disconnectOptions_initializer5 with the two missing
fields (onSuccelss5 and onFailure5) set to NULL. Update the previous
initializer and set the struct_version to 0 as it should not be used for
MQTTv5 anyway.

Signed-off-by: Julien Grossholtz <julien.grossholtz@openest.io>