Issue:With MQTTCLIENT_PERSISTENCE_DEFAULT the paho code crashes for MQTTVERSION_5 #1108
For MQTTVERSION_5 for SUBSCRIBE command qoss details was not added in persistent folder.
Also at time of retrieving for SUBSCRIBE & UNSUBSCRIBE command ptr value was not update properly for MQTTProperties value in the buffer.
Because of which MQTTProperties_read() would not return 1 and qcommand ends being free and later down the call stack crash would happen.

This fix covers above two issue.